backend/requirements.txt
Django==2.2.10
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-templated-mail==1.1.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
git+http://github.com/NiyazNz/django-rest-framework-simplejwt@jwt_cookie#egg=djangorestframework-simplejwt
djoser==2.0.3
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
